requirements.txt
PySide6
numpy
xraylarch
